refactor: split editor.js into focused manager classes (closes #37)
